+++ R/FBM-copy.R
@@ -1,14 +1,15 @@
 ################################################################################
 
-#' Copy a Filebacked Big Matrix
+#' Copy as a Filebacked Big Matrix
 #'
-#' Copy a Filebacked Big Matrix with possible subsetting.
+#' Deep copy of a Filebacked Big Matrix with possible subsetting.
+#' This should also work for any matrix-like object.
 #'
 #' @param X Could be any matrix-like object.
 #' @inheritParams bigstatsr-package
 #' @inheritParams FBM
 #'
-#' @return A copy of the [FBM][FBM-class].
+#' @return A copy of `X` as a new [FBM][FBM-class] object.
 #' @export
 #'
 #' @examples
@@ -18,7 +19,7 @@
 #' X2[]
 #'
 #' mat <- matrix(101:200, 10)
-#' X3 <- big_copy(mat, type = "double")
+#' X3 <- big_copy(mat, type = "double")  # as_FBM() would be faster here
 #' X3[]
 #'
 #' X.code <- big_attachExtdata()
@@ -31,7 +32,8 @@ big_copy <- function(X, ind.row = rows_along(X),
                      ind.col = cols_along(X),
                      type = typeof(X),
                      backingfile = tempfile(),
-                     block.size = block_size(length(ind.row))) {
+                     block.size = block_size(length(ind.row)),
+                     is_read_only = FALSE) {
 
   if (inherits(X, "FBM.code256") && type == "unsigned char") {
     args <- as.list(environment())
@@ -45,7 +47,8 @@ big_copy <- function(X, ind.row = rows_along(X),
     ncol = length(ind.col),
     init = NULL,
     type = type,
-    backingfile = backingfile
+    backingfile = backingfile,
+    is_read_only = FALSE
   )
 
   # Don't write in parallel
@@ -55,6 +58,7 @@ big_copy <- function(X, ind.row = rows_along(X),
   }, a.combine = 'c', ind = seq_along(ind.col), block.size = block.size,
   X2 = res, ind.row = ind.row, ind.col = ind.col)
 
+  res$is_read_only <- is_read_only
   res
 }