+++ R/biglasso.R
@@ -15,13 +15,13 @@ null_pred <- function(var0, y, base, family) {
 summaries <- function(X, y_diff.train, ind.train, ind.col, ind.sets, K,
                       covar.train = matrix(0, length(ind.train), 0)) {
 
-  summ <- bigsummaries(X, ind.train, ind.col, covar.train, y_diff.train,
-                       ind.sets, K)
+  all_sums <-
+    bigsummaries(X, ind.train, ind.col, covar.train, y_diff.train, ind.sets, K)
 
-  all <- colSums(summ)
-  SUM_X  <- sweep(-matrix(summ[, , 1], K), 2, all[, 1], '+')
-  SUM_XX <- sweep(-matrix(summ[, , 2], K), 2, all[, 2], '+')
-  SUM_XY <- sweep(-matrix(summ[, , 3], K), 2, all[, 3], '+')
+  SUM_X  <- sweep(-all_sums$sumX,  2, colSums(all_sums$sumX),  '+')
+  SUM_XX <- sweep(-all_sums$sumXX, 2, colSums(all_sums$sumXX), '+')
+  SUM_XY <- sweep(-all_sums$sumXY, 2, colSums(all_sums$sumXY), '+')
+
   SUM_Y <- tapply(seq_along(ind.sets), factor(ind.sets, levels = 1:K),
                   function(ind) sum(y_diff.train[-ind]))
 
@@ -174,7 +174,8 @@ COPY_biglasso_part <- function(X, y.train, ind.train, ind.col, covar.train,
 #'   which set each index of the training set is in. Default randomly assigns
 #'   these values but it can be useful to set this vector for reproducibility,
 #'   or if you want to refine the grid-search over `alphas` using the same sets.
-#' @param warn Deprecated. Now return the reason of completion as `$message`.
+#' @param warn Whether to warn if some models may not have reached a minimum.
+#'   Default is `TRUE`.
 #' @param return.all Deprecated. Now always return all models.
 #' @param nlam.min Minimum number of lambda values to investigate. Default is `50`.
 #' @param n.abort Number of lambda values for which prediction on the validation
@@ -188,6 +189,8 @@ COPY_biglasso_part <- function(X, y.train, ind.train, ind.col, covar.train,
 #'   penalization if some coefficients are thought to be more likely than others
 #'   to be in the model. Setting SOME to 0 allows to have unpenalized coefficients.
 #' @param pf.covar Same as `pf.X`, but for `covar.train`.
+#'   You might want to set some to 0 as variables with large effects can mask
+#'   small effects in penalized regression.
 #'
 #' @keywords internal
 #'
@@ -208,10 +211,9 @@ COPY_biglasso_main <- function(X, y.train, ind.train, ind.col, covar.train,
                                dfmax = 50e3,
                                lambda.min = `if`(n > p, .0001, .001),
                                return.all = FALSE,
-                               warn = FALSE,
+                               warn = TRUE,
                                ncores = 1) {
 
-  if (!missing(warn)) warning2("Parameter 'warn' is deprecated.")
   if (!missing(return.all)) warning2("Parameter 'return.all' is deprecated.")
   if (!missing(lambda.min)) {
     lambda.min.ratio <- lambda.min
@@ -332,7 +334,7 @@ COPY_biglasso_main <- function(X, y.train, ind.train, ind.col, covar.train,
              `if`(nb_novar == 1, "", "s"), `if`(nb_novar == 1, "has", "have"),
              "Access remaining columns with 'attr(<object>, \"ind.col\")'.")
 
-  structure(
+  res <- structure(
     cross.res,
     class = "big_sp_list",
     family = family,
@@ -342,6 +344,12 @@ COPY_biglasso_main <- function(X, y.train, ind.train, ind.col, covar.train,
     pf = pf.keep,
     base = base.train0
   )
+
+  if (warn && !all(summary(res)$all_conv))
+    warning2("Some models may not have reached a minimum; %s",
+             "check with summary() and plot().")
+
+  res
 }
 
 ################################################################################
@@ -373,7 +381,7 @@ COPY_biglasso_main <- function(X, y.train, ind.train, ind.col, covar.train,
 #'
 #' @inheritParams bigstatsr-package
 #' @inheritParams COPY_biglasso_main
-#' @inheritDotParams COPY_biglasso_main lambda.min.ratio eps max.iter warn return.all
+#' @inheritDotParams COPY_biglasso_main lambda.min.ratio eps max.iter return.all
 #'
 #' @return Return an object of class `big_sp_list` (a list of `length(alphas)`
 #'   x `K`) that has 3 methods `predict`, `summary` and `plot`.
@@ -412,6 +420,7 @@ big_spLinReg <- function(X, y.train,
                          nlam.min = 50,
                          n.abort = 10,
                          dfmax = 50e3,
+                         warn = TRUE,
                          ncores = 1,
                          ...) {
 
@@ -429,7 +438,7 @@ big_spLinReg <- function(X, y.train,
 #'
 #' @inheritParams bigstatsr-package
 #' @inheritParams COPY_biglasso_main
-#' @inheritDotParams COPY_biglasso_main lambda.min.ratio eps max.iter warn return.all
+#' @inheritDotParams COPY_biglasso_main lambda.min.ratio eps max.iter return.all
 #'
 #' @inherit big_spLinReg return description details seealso references
 #'
@@ -450,6 +459,7 @@ big_spLogReg <- function(X, y01.train,
                          nlam.min = 50,
                          n.abort = 10,
                          dfmax = 50e3,
+                         warn = TRUE,
                          ncores = 1,
                          ...) {