+++ inst/include/bigstatsr/BMAcc.h
@@ -8,12 +8,14 @@
 using namespace Rcpp;
 using std::size_t;
 
+#define NA_FLOAT __FLT_MIN__
+
 /******************************************************************************/
 
 template <typename T>
-class BMAcc {
+class BMAcc_RW {
 public:
-  BMAcc(FBM * xpBM) {
+  BMAcc_RW(FBM_RW * xpBM) {
     _pMat = static_cast<T*>(xpBM->matrix());
     _nrow = xpBM->nrow();
     _ncol = xpBM->ncol();
@@ -40,34 +42,59 @@ protected:
 /******************************************************************************/
 
 template <typename T>
-class VecBMAcc : public BMAcc<T> {
+class SubBMAcc_RW : public BMAcc_RW<T> {
 public:
-  VecBMAcc(FBM * xpBM,
-           const NumericVector& elem_ind)
-    : BMAcc<T>(xpBM) {
-
-      size_t ind, k;
-      size_t LIM = xpBM->nrow() * xpBM->ncol();
-      size_t n = elem_ind.size();
-      std::vector<size_t> elem_ind2(n);
-      for (k = 0; k < n; k++) {
-        ind = static_cast<size_t>(elem_ind[k]);
-        myassert_bounds(ind, LIM);
-        elem_ind2[k] = ind;
-      }
-      _elem_ind = elem_ind2;
+  SubBMAcc_RW(FBM_RW * xpBM,
+              const IntegerVector& row_ind,
+              const IntegerVector& col_ind,
+              int sub = 0)
+    : BMAcc_RW<T>(xpBM) {
+      _row_ind = vec_int_to_size(row_ind, xpBM->nrow(), sub);
+      _col_ind = vec_int_to_size(col_ind, xpBM->ncol(), sub);
     }
 
-  inline T& operator[](size_t k) {
-    // https://stackoverflow.com/a/7076312/6103040
-    return this->_pMat[_elem_ind[k]];
+  inline T& operator()(size_t i, size_t j) {
+    // https://stackoverflow.com/a/32087373/6103040
+    return BMAcc_RW<T>::operator()(_row_ind[i], _col_ind[j]);
   }
 
-  size_t size() const { return _elem_ind.size(); }
-  size_t nelem() const { return this->size(); }
+  // WARNING: operator[] is not redefined
+
+  size_t nrow() const { return _row_ind.size(); }
+  size_t ncol() const { return _col_ind.size(); }
 
 protected:
-  std::vector<size_t> _elem_ind;
+  std::vector<size_t> _row_ind;
+  std::vector<size_t> _col_ind;
+};
+
+/******************************************************************************/
+
+template <typename T>
+class BMAcc {
+public:
+  BMAcc(FBM * xpBM) {
+    _pMat = static_cast<const T*>(xpBM->matrix());
+    _nrow = xpBM->nrow();
+    _ncol = xpBM->ncol();
+  }
+
+  inline T operator()(size_t i, size_t j) {
+    return _pMat[i + j * _nrow];
+  }
+
+  inline T operator[](size_t k) {
+    return _pMat[k];
+  }
+
+  size_t nrow() const { return _nrow; }
+  size_t ncol() const { return _ncol; }
+  size_t size() const { return _nrow * _ncol; }
+
+protected:
+  const T* _pMat;
+  size_t _nrow;
+  size_t _ncol;
 };
 
 /******************************************************************************/
@@ -77,38 +104,19 @@ class SubBMAcc : public BMAcc<T> {
 public:
   SubBMAcc(FBM * xpBM,
            const IntegerVector& row_ind,
-           const IntegerVector& col_ind)
+           const IntegerVector& col_ind,
+           int sub = 0)
     : BMAcc<T>(xpBM) {
-
-      size_t ind, i, j;
-
-      size_t LIM_N = xpBM->nrow();
-      size_t n = row_ind.size();
-      std::vector<size_t> row_ind2(n);
-      for (i = 0; i < n; i++) {
-        ind = static_cast<size_t>(row_ind[i]);
-        myassert_bounds(ind, LIM_N);
-        row_ind2[i] = ind;
-      }
-      _row_ind = row_ind2;
-
-      size_t LIM_M = xpBM->ncol();
-      size_t m = col_ind.size();
-      std::vector<size_t> col_ind2(m);
-      for (j = 0; j < m; j++) {
-        ind = static_cast<size_t>(col_ind[j]);
-        myassert_bounds(ind, LIM_M);
-        col_ind2[j] = ind;
-      }
-      _col_ind = col_ind2;
+      _row_ind = vec_int_to_size(row_ind, xpBM->nrow(), sub);
+      _col_ind = vec_int_to_size(col_ind, xpBM->ncol(), sub);
     }
 
-  inline T& operator()(size_t i, size_t j) {
+  inline T operator()(size_t i, size_t j) {
     // https://stackoverflow.com/a/32087373/6103040
     return BMAcc<T>::operator()(_row_ind[i], _col_ind[j]);
-    // https://stackoverflow.com/a/7076312/6103040
-    // return this->_pMat[_row_ind[i] + _col_ind[j] * this->_nrow];
   }
+
+  // WARNING: operator[] is not redefined
 
   size_t nrow() const { return _row_ind.size(); }
   size_t ncol() const { return _col_ind.size(); }
@@ -116,6 +124,49 @@ public:
 protected:
   std::vector<size_t> _row_ind;
   std::vector<size_t> _col_ind;
+};
+
+/******************************************************************************/
+
+// For biglasso
+template<typename T>
+class SubMatCovAcc : public SubBMAcc<T> {
+public:
+  SubMatCovAcc(FBM * xpBM,
+               const IntegerVector& row_ind,
+               const IntegerVector& col_ind,
+               const NumericMatrix& covar,
+               int sub = 0)
+    : SubBMAcc<T>(xpBM, row_ind, col_ind, sub) {
+
+      _ncolsub = col_ind.size();
+
+      if (covar.nrow() != 0) {
+        myassert_size(row_ind.size(), covar.nrow());
+        _ncoladd = covar.ncol();
+        _covar = covar;
+      }  else {
+        _ncoladd = 0;
+      }
+    }
+
+  inline double operator() (size_t i, size_t j) {
+    int j2 = j - _ncolsub;
+    if (j2 < 0) {
+      // https://stackoverflow.com/a/32087373/6103040
+      return SubBMAcc<T>::operator()(i, j);
+    } else {
+      return _covar(i, j2);
+    }
+  }
+
+  size_t nrow() const { return this->_row_ind.size(); }
+  size_t ncol() const { return _ncolsub + _ncoladd; }
+
+protected:
+  size_t _ncolsub;
+  size_t _ncoladd;
+  NumericMatrix _covar;
 };
 
 /******************************************************************************/